refactor: use custom auth insteam of moleculer's
There's an issue with the authorize method and moleculer's tracking
where any non-trivial promise that is executed causes a second
context to be created that never finishes. This means, that any
`broker.call` or raw database call that is executed causes the
endpoint to hang indefinitely. This is only an issue on REST
endpoints and not websocket topics.